+++ src/App.tsx
@@ -9,8 +9,6 @@ import {
   Stack,
   Typography,
 } from '@mui/material';
-import { useSnackbar } from 'notistack';
-import { useState } from 'react';
 
 import { CalendarView } from './components/CalendarView';
 import { EventFormPanel } from './components/EventFormPanel';
@@ -21,83 +19,29 @@ import { useEventForm } from './hooks/useEventForm';
 import { useEventOperations } from './hooks/useEventOperations';
 import { useFilteredEvents } from './hooks/useFilteredEvents';
 import { useNotifications } from './hooks/useNotifications';
-import { Event, EventForm } from './types';
-import { findOverlappingEvents } from './utils/eventOverlap';
+import { useSubmitHandler } from './hooks/useSubmitHandler';
 
 function App() {
   const formProps = useEventForm();
-  const {
-    editingEvent,
-    resetForm,
-    title,
-    date,
-    startTime,
-    endTime,
-    description,
-    location,
-    category,
-    isRepeating,
-    repeatType,
-    repeatInterval,
-    repeatEndDate,
-    notificationTime,
-    startTimeError,
-    endTimeError,
-  } = formProps;
-
-  const { events, saveEvent, deleteEvent } = useEventOperations(Boolean(editingEvent), () => {
+  const eventOperations = useEventOperations(Boolean(formProps.editingEvent), () => {
     formProps.setEditingEvent(null);
-    resetForm();
+    formProps.resetForm();
   });
-
-  const { notifications, notifiedEvents, removeNotification } = useNotifications(events);
   const calendarViewProps = useCalendarView();
-  const { filteredEvents, searchTerm, setSearchTerm } = useFilteredEvents(
-    events,
+  const notificationProps = useNotifications(eventOperations.events);
+  const filteredEventsProps = useFilteredEvents(
+    eventOperations.events,
     calendarViewProps.currentDate,
     calendarViewProps.view
   );
 
-  const [isOverlapDialogOpen, setIsOverlapDialogOpen] = useState(false);
-  const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
-  const { enqueueSnackbar } = useSnackbar();
-
-  const addOrUpdateEvent = async () => {
-    if (!title || !date || !startTime || !endTime) {
-      enqueueSnackbar('필수 정보를 모두 입력해주세요.', { variant: 'error' });
-      return;
-    }
-    if (startTimeError || endTimeError) {
-      enqueueSnackbar('시간 설정을 확인해주세요.', { variant: 'error' });
-      return;
-    }
-
-    const eventData: Event | EventForm = {
-      id: editingEvent ? editingEvent.id : undefined,
-      title,
-      date,
-      startTime,
-      endTime,
-      description,
-      location,
-      category,
-      repeat: {
-        type: isRepeating ? repeatType : 'none',
-        interval: repeatInterval,
-        endDate: repeatEndDate || undefined,
-      },
-      notificationTime,
-    };
-
-    const overlapping = findOverlappingEvents(eventData, events);
-    if (overlapping.length > 0) {
-      setOverlappingEvents(overlapping);
-      setIsOverlapDialogOpen(true);
-    } else {
-      await saveEvent(eventData);
-      resetForm();
-    }
-  };
+  const {
+    addOrUpdateEvent,
+    isOverlapDialogOpen,
+    overlappingEvents,
+    confirmOverlap,
+    cancelOverlap,
+  } = useSubmitHandler(formProps, eventOperations);
 
   return (
     <Box sx={{ width: '100%', height: '100vh', margin: 'auto', p: 5 }}>
@@ -106,24 +50,22 @@ function App() {
         <CalendarView
           viewProps={{
             ...calendarViewProps,
-            filteredEvents,
-            notifiedEvents,
+            ...filteredEventsProps,
+            ...notificationProps,
             editEvent: formProps.editEvent,
           }}
         />
         <EventListPanel
           listProps={{
-            searchTerm,
-            setSearchTerm,
-            filteredEvents,
-            notifiedEvents,
+            ...filteredEventsProps,
+            ...notificationProps,
             editEvent: formProps.editEvent,
-            deleteEvent,
+            deleteEvent: eventOperations.deleteEvent,
           }}
         />
       </Stack>
 
-      <Dialog open={isOverlapDialogOpen} onClose={() => setIsOverlapDialogOpen(false)}>
+      <Dialog open={isOverlapDialogOpen} onClose={cancelOverlap}>
         <DialogTitle>일정 겹침 경고</DialogTitle>
         <DialogContent>
           <DialogContentText>
@@ -137,36 +79,17 @@ function App() {
           </DialogContentText>
         </DialogContent>
         <DialogActions>
-          <Button onClick={() => setIsOverlapDialogOpen(false)}>취소</Button>
-          <Button
-            color="error"
-            onClick={async () => {
-              setIsOverlapDialogOpen(false);
-              await saveEvent({
-                id: editingEvent ? editingEvent.id : undefined,
-                title,
-                date,
-                startTime,
-                endTime,
-                description,
-                location,
-                category,
-                repeat: {
-                  type: isRepeating ? repeatType : 'none',
-                  interval: repeatInterval,
-                  endDate: repeatEndDate || undefined,
-                },
-                notificationTime,
-              });
-              resetForm();
-            }}
-          >
+          <Button onClick={cancelOverlap}>취소</Button>
+          <Button color="error" onClick={confirmOverlap}>
             계속 진행
           </Button>
         </DialogActions>
       </Dialog>
 
-      <NotificationPanel notifications={notifications} onDismiss={removeNotification} />
+      <NotificationPanel
+        notifications={notificationProps.notifications}
+        onDismiss={notificationProps.removeNotification}
+      />
     </Box>
   );
 }
